+++ src/extensions.js
@@ -36,7 +36,7 @@ TableFrameMorph, console, Morph, MenuMorph*/
 
 /*jshint esversion: 11, bitwise: false*/
 
-modules.extensions = '2026-August-03';
+modules.extensions = '2026-Septembr-07';
 
 // Global stuff
 
@@ -1944,6 +1944,20 @@ SnapExtensions.primitives.set(
             return;
         }
         return new List([rect.left(), rect.top(), rect.right(), rect.bottom()]);
+    }
+);
+
+SnapExtensions.primitives.set(
+    'scn_label(txt)',
+    function (txt, proc) {
+        var stage = this.parentThatIsA(StageMorph),
+            dlg;
+
+        if (!stage.tutorialMode) {return ''; }
+        dlg = stage.parentThatIsA(DialogBoxMorph);
+        dlg.labelString = txt;
+        dlg.createLabel();
+        dlg.fixLayout();
     }
 );